reconfiguring files to be implemented into the rootReducer and types.js files
